Localized helper methods and taxonomy loading to article-feed
