+++ libs/utils/utils.js
@@ -1,5 +1,3 @@
-import * as taxonomyLibrary from '../scripts/taxonomy.js';
-
 const PROJECT_NAME = 'milo--adobecom';
 const PRODUCTION_DOMAINS = ['milo.adobe.com'];
 const MILO_TEMPLATES = [
@@ -82,9 +80,9 @@ function getEnv(conf) {
   if (host.includes('localhost:')) return { ...ENVS.local, consumer: conf.local };
   /* c8 ignore start */
   if (host.includes('hlx.page')
-   || host.includes('hlx.live')
-   || host.includes('stage.adobe')
-   || host.includes('corp.adobe')) {
+    || host.includes('hlx.live')
+    || host.includes('stage.adobe')
+    || host.includes('corp.adobe')) {
     return { ...ENVS.stage, consumer: conf.stage };
   }
   return { ...ENVS.prod, consumer: conf.prod };
@@ -243,7 +241,7 @@ export async function loadBlock(block) {
         await init(block);
       } catch (err) {
         // eslint-disable-next-line no-console
-        console.warn(`Failed loading ${name}`, err);
+        console.log(`Failed loading ${name}`, err);
         const config = getConfig();
         if (config.env.name !== 'prod') {
           block.dataset.failed = 'true';
@@ -427,6 +425,76 @@ export async function loadDeferred(area) {
   }
 }
 
+/**
+* Load the Privacy library
+*/
+function loadPrivacy() {
+  // Configure Privacy
+  window.fedsConfig = {
+    privacy: {
+      otDomainId: '7a5eb705-95ed-4cc4-a11d-0cc5760e93db',
+      footerLinkSelector: '[href="https://www.adobe.com/#openPrivacy"]',
+    },
+  };
+
+  const env = getConfig().env.name === 'prod' ? '' : 'stage.';
+  loadScript(`https://www.${env}adobe.com/etc.clientlibs/globalnav/clientlibs/base/privacy-standalone.js`);
+}
+
+export async function loadArea(area = document) {
+  const config = getConfig();
+  const isDoc = area === document;
+
+  if (isDoc) {
+    decorateHeader();
+  }
+
+  const sections = decorateSections(area, isDoc);
+  // eslint-disable-next-line no-restricted-syntax
+  for (const section of sections) {
+    const loaded = section.blocks.map((block) => loadBlock(block));
+
+    // Only move on to the next section when all blocks are loaded.
+    // eslint-disable-next-line no-await-in-loop
+    await Promise.all(loaded);
+
+    // Post LCP operations.
+    if (isDoc && section.el.dataset.idx === '0') { loadPostLCP(config); }
+
+    // Show the section when all blocks inside are done.
+    delete section.el.dataset.status;
+    delete section.el.dataset.idx;
+  }
+
+  // Post section loading on document
+  if (isDoc) {
+    loadFooter();
+    const { default: loadFavIcon } = await import('./favicon.js');
+    loadFavIcon(createTag, getConfig(), getMetadata);
+  }
+
+  // Load everything that can be deferred until after all blocks load.
+  await loadDeferred(area);
+}
+
+/**
+ * Load everything that impacts performance later.
+ */
+export function loadDelayed(delay = 3000) {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      loadPrivacy();
+      if (getMetadata('interlinks') === 'on') {
+        import('../features/interlinks.js').then((mod) => {
+          resolve(mod);
+        });
+      } else {
+        resolve(null);
+      }
+    }, delay);
+  });
+}
+
 export function utf8ToB64(str) {
   return window.btoa(unescape(encodeURIComponent(str)));
 }
@@ -500,219 +568,4 @@ export function createIntersectionObserver({ el, callback, once = true, options 
   }, options);
   io.observe(el);
   return io;
-}
-
-export const getLocaleIetf = () => getConfig().locale?.ietf;
-
-/**
- * Returns the language dependent root path
- * @returns {string} The computed root path
- */
-export const getPrefix = () => getConfig().locale?.prefix;
-
-export const emptyDiv = () => document.createElement('div');
-
-/**
- * fetches the string variables.
- * @returns {object} localized variables
- */
-
-export async function fetchPlaceholders() {
-  if (!window.placeholders) {
-    const resp = await fetch(`/${getPrefix()}/placeholders.json`);
-    const json = await resp.json();
-    window.placeholders = {};
-    json.data.forEach((placeholder) => {
-      window.placeholders[placeholder.Key] = placeholder.Text;
-    });
-  }
-  return window.placeholders;
-}
-
-export function stamp(message) {
-  if (window.name.includes('performance')) {
-    // eslint-disable-next-line no-console
-    console.warn(`${new Date() - performance.timeOrigin}:${message}`);
-  }
-}
-
-/*
- *
- * Taxonomy Utils
- *
-*/
-
-let taxonomyModule;
-
-export function getTaxonomyModule() {
-  return taxonomyModule;
-}
-
-/**
- * For the given list of topics, returns the corresponding computed taxonomy:
- * - category: main topic
- * - topics: tags as an array
- * - visibleTopics: list of visible topics, including parents
- * - allTopics: list of all topics, including parents
- * @param {Array} topics List of topics
- * @returns {Object} Taxonomy object
- */
-function computeTaxonomyFromTopics(taxonomy, topics, path) {
-  // no topics: default to a randomly choosen category
-  const category = topics?.length > 0 ? topics[0] : 'news';
-
-  if (taxonomy) {
-    const allTopics = [];
-    const visibleTopics = [];
-    // if taxonomy loaded, we can compute more
-    topics?.forEach((tag) => {
-      const tax = taxonomy.get(tag);
-      if (tax) {
-        if (!allTopics.includes(tag) && !tax.skipMeta) {
-          allTopics.push(tag);
-          if (tax.isUFT) visibleTopics.push(tag);
-          const parents = taxonomy.getParents(tag);
-          if (parents) {
-            parents.forEach((parent) => {
-              const ptax = taxonomy.get(parent);
-              if (!allTopics.includes(parent)) {
-                allTopics.push(parent);
-                if (ptax.isUFT) visibleTopics.push(parent);
-              }
-            });
-          }
-        }
-      } else {
-        // eslint-disable-next-line no-console
-        console.warn(`Unknown topic in tags list: ${tag} ${path ? `on page ${path}` : '(current page)'}`);
-      }
-    });
-    return { category, topics, visibleTopics, allTopics };
-  }
-  return { category, topics };
-}
-
-export async function loadTaxonomy() {
-  taxonomyModule = await taxonomyLibrary.default(getPrefix());
-  if (taxonomyModule) {
-    // taxonomy loaded, post loading adjustments
-    // fix the links which have been created before the taxonomy has been loaded
-    // (pre lcp or in lcp block).
-    document.querySelectorAll('[data-topic-link]').forEach((a) => {
-      const topic = a.dataset.topicLink;
-      const tax = taxonomyModule.get(topic);
-      if (tax) {
-        a.href = tax.link;
-      } else {
-        // eslint-disable-next-line no-console
-        console.warn(`Trying to get a link for an unknown topic: ${topic} (current page)`);
-        a.href = '#';
-      }
-      delete a.dataset.topicLink;
-    });
-
-    // adjust meta article:tag
-
-    const currentTags = getMetadata('article:tag', true) || [];
-    const articleTax = computeTaxonomyFromTopics(taxonomyModule, currentTags);
-
-    const allTopics = articleTax.allTopics || [];
-    allTopics.forEach((topic) => {
-      if (!currentTags.includes(topic)) {
-        // computed topic (parent...) is not in meta -> add it
-        const newMetaTag = document.createElement('meta');
-        newMetaTag.setAttribute('property', 'article:tag');
-        newMetaTag.setAttribute('content', topic);
-        document.head.append(newMetaTag);
-      }
-    });
-
-    currentTags.forEach((tag) => {
-      const tax = taxonomyModule.get(tag);
-      if (tax && tax.skipMeta) {
-        // if skipMeta, remove from meta "article:tag"
-        const meta = document.querySelector(`[property="article:tag"][content="${tag}"]`);
-        if (meta) {
-          meta.remove();
-        }
-        // but add as meta with name
-        const newMetaTag = document.createElement('meta');
-        newMetaTag.setAttribute('name', tag);
-        newMetaTag.setAttribute('content', 'true');
-        document.head.append(newMetaTag);
-      }
-    });
-  }
-}
-
-/**
-* Load the Privacy library
-*/
-function loadPrivacy() {
-  // Configure Privacy
-  window.fedsConfig = {
-    privacy: {
-      otDomainId: '7a5eb705-95ed-4cc4-a11d-0cc5760e93db',
-      footerLinkSelector: '[href="https://www.adobe.com/#openPrivacy"]',
-    },
-  };
-
-  const env = getConfig().env.name === 'prod' ? '' : 'stage.';
-  loadScript(`https://www.${env}adobe.com/etc.clientlibs/globalnav/clientlibs/base/privacy-standalone.js`);
-}
-
-export async function loadArea(area = document) {
-  const config = getConfig();
-  const isDoc = area === document;
-
-  if (isDoc) {
-    decorateHeader();
-  }
-
-  await loadTaxonomy();
-
-  const sections = decorateSections(area, isDoc);
-  // eslint-disable-next-line no-restricted-syntax
-  for (const section of sections) {
-    const loaded = section.blocks.map((block) => loadBlock(block));
-
-    // Only move on to the next section when all blocks are loaded.
-    // eslint-disable-next-line no-await-in-loop
-    await Promise.all(loaded);
-
-    // Post LCP operations.
-    if (isDoc && section.el.dataset.idx === '0') { loadPostLCP(config); }
-
-    // Show the section when all blocks inside are done.
-    delete section.el.dataset.status;
-    delete section.el.dataset.idx;
-  }
-
-  // Post section loading on document
-  if (isDoc) {
-    loadFooter();
-    const { default: loadFavIcon } = await import('./favicon.js');
-    loadFavIcon(createTag, getConfig(), getMetadata);
-  }
-
-  // Load everything that can be deferred until after all blocks load.
-  await loadDeferred(area);
-}
-
-/**
- * Load everything that impacts performance later.
- */
-export function loadDelayed(delay = 3000) {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      loadPrivacy();
-      if (getMetadata('interlinks') === 'on') {
-        import('../features/interlinks.js').then((mod) => {
-          resolve(mod);
-        });
-      } else {
-        resolve(null);
-      }
-    }, delay);
-  });
 }
